feat: 4.23* bloglist expansion step11, adding userExtractor middleware

+++ Part4/bloglist/controllers/blogs.js
@@ -12,24 +12,23 @@ blogsRouter.get("/", async (request, response) => {
 blogsRouter.post("/", async (request, response, next) => {
   const body = request.body;
 
-  const token = request.token;
-  const decodedToken = jwt.verify(token, process.env.SECRET);
-
   if (!body.title || !body.url) {
     return response.status(400).end();
   } else if (!body.likes) {
     body.likes = 0;
   }
 
-  //fe
-  const user = await User.findById(decodedToken.id);
+  const user = request.user;
+
+  if (!user) {
+    return response.status(401).json({ error: "token is missing or invalid" });
+  }
 
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes,
-    //fe
     user: user._id,
   });
 
@@ -42,9 +41,11 @@ blogsRouter.post("/", async (request, response, next) => {
 blogsRouter.delete("/:id", async (request, response) => {
   const blog = await Blog.findById(request.params.id);
 
-  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!blog) {
+    return response.status(204).end();
+  }
 
-  const user = await User.findById(decodedToken.id);
+  const user = request.user;
 
   if (user.id.toString() === blog.user.toString()) {
     await Blog.findByIdAndRemove(request.params.id);
